+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - #81: added integration tests
 - #32: added `while` keyword
 - #29: added ternary operator
+- #55: added post increment and decrement operators `++` and `--`
 
 ### Changed
 - #3: prefix private class members with `m_`